Empty command execution + test
